+++ src/api/providers/rest/routes.go
@@ -1,8 +1,10 @@
 package rest
 
 const (
-	GetHealthRoute                         = "/health"
+	HealthRoute                            = "/health"
+	AboutRoute                             = "/about"
 	AccountLoginRoute                      = "/login"
+	AccountLogoutRoute                     = "/logout"
 	AddAccountRoute                        = "/accounts"
 	GetAccountsRoute                       = "/accounts"
 	DeleteAccountRoute                     = "/accounts"