+++ src/app/models/account.go
@@ -7,30 +7,30 @@ import (
 )
 
 type Account struct {
-	Id     string
-	AuthId string
-	AuthPw string
+	Id       string
+	Username string
+	Password string
 }
 
-func NewValidAccount(authId, authPw string) (*Account, error) {
-	if authId == "" {
-		return nil, errors.New("authId is empty")
+func NewValidAccount(username, password string) (*Account, error) {
+	if username == "" {
+		return nil, errors.New("username is empty")
 	}
 
-	if authPw == "" {
-		return nil, errors.New("authPw is empty")
+	if password == "" {
+		return nil, errors.New("password is empty")
 	}
 
 	return &Account{
-		AuthId: authId,
-		AuthPw: authPw,
+		Username: username,
+		Password: password,
 	}, nil
 }
 
 func AcccountDBModelToAccount(accountDBModel models.AccountDBModel) Account {
 	return Account{
-		Id:     accountDBModel.Id,
-		AuthId: accountDBModel.AuthId,
-		AuthPw: accountDBModel.AuthPw,
+		Id:       accountDBModel.Id,
+		Username: accountDBModel.Username,
+		Password: accountDBModel.Password,
 	}
 }