+++ src/database/models/account.go
@@ -1,7 +1,7 @@
 package models
 
 type AccountDBModel struct {
-	Id     string
-	AuthId string
-	AuthPw string
+	Id       string
+	Username string
+	Password string
 }