+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -->
 
 ## [Unreleased]
+## [v0.5.1] - 2021-10-11
+### Added
+- Account can now logout
+- About Route
+- Version for the about route gets passed by the github action
+### Changed
+- Changed `auth_id` and `auth_pw` to `username` and `password`
+### Fixed
+- While creating a new account, a user error was referenced as an internal error
 ## [v0.5.0] - 2021-10-06
 ### Added
 - More debug logging messages
@@ -48,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (Un)register to the substitution updater which automatically sends updates on new substitutions
 
 [Unreleased]: https://github.com/Dattito/purrmannplus-backend/tree/dev
+[v0.5.1]: https://github.com/Dattito/purrmannplus-backend/compare/v0.5.0...v0.5.1
 [v0.5.0]: https://github.com/Dattito/purrmannplus-backend/compare/v0.4.1...v0.5.0
 [v0.4.1]: https://github.com/Dattito/purrmannplus-backend/compare/v0.4.0...v0.4.1
 [v0.4.0]: https://github.com/Dattito/purrmannplus-backend/compare/v0.3.0...v0.4.0